feat(qui-cli.core)!: core is module plugin registrar
`core` now checks to see if plugin dependencies are
already installed before installing the plugin, throwing
an error if they are missing.

Added `shelljs` out of effeteness and a desire to be
able to commitlint for console.log() calls in my code.

BREAKING: core as a Plugin.Registrar now requires all
plugins to adhere to the module plugin definition.