feat(qui-cli.plugin)!: plugin modules and Plugin.Registrar
Re-designed plugins (close #15) to be written as ESM modules, rather than
TypeScript classes. This better models the desired singleton
implementation, allowing dependencies to be managed by the package
manager.

BREAKING: All plugins that are implemented as extensions of the
Plugin.Base class are broken and need to be reimplemented as ESM
modules.