feat(qui-cli)!: reimplemented as ESM modules
While `cli` continues to follow the same basic structure as in the past,
hanging the built-in plugins off of `cli` as properties, all plugin
management is relegated to Core and Plugin.Registrar. `cli` now exists only
as a convenience package to auto-import all of the built-in plugins at once.

Progress is temporarily removed until it can be reimplemented as an ESM
plugin.

BREAKING: `cli` has changed its behavior, now requiring seprate
`configure()` and `init()` calls to configure Jackspeaks’s behavior and pass
in expected arguments, respectively.