fix: use a K.I.S.S. approach to determining package conflicts
The concern is that multiple implementations (and thus multiple separate modules)
representing the same plugin could be loaded, thus inhibiting interoperability
between plugins. (If Plugin A is interacting with v1 of Plugin B, but Plugin C is
interacting with v2 of Plugin B, the potential then assumptions about the behavior
of Plugin B are no longer safe!

Distributed plugins should include their dependencies as `peerDependencies`
to ensure that all plugins are hoisted to the same level and consistent
versions/modules for all plugins are used throughout.